requirements/common.txt
# keras-rl2==1.0.5
gym>=0.21
matplotlib>=3.5.3
stable-baselines3>=1.6.2
torch-tb-profiler>=0.4.0
python-dotenv>=0.21.0
pybind11
python-dotenv
tqdm
readchar
japanize_matplotlib
